refactor: Update Dataspace routines and Group array access (WIP)
Update Group.array to be None instead of raising an error if no Array is
defined (this is what docstrings states anyway). Add routines and
clean-up Dataspacrefactor: Update Dataspace routines and Group array
access (WIP)

Update Group.array to be None instead of raising an error if no Array is
defined (this is what docstrings states anyway). Add routines and
clean-up Dataspace.