Add support for converting a NetCDF group where attributes are stored in separate arrays (#21)
Features: 

1. Support auto-converting from NetCDF to TileDB by either combining all
attributes with the same dimensions into a single array or by creating
an array for each attribute.

2. Split `tiles` parameter into `tiles_by_var` and `tiles_by_dims`. The natural 
way to support tiles when splitting variables into a separate array each is
to set the tiles by specifying the variable name, but when collecting
variables with the same underlying dimensions, it is easier to specifying the
tiles by the dimension names. This adds support for both options (or a
combination of the two).