Changes xarray backend to save timestamp at time of opening (#120)
This PR updates the TileDB-xarray backend so that it returns the data from the timestamp specified when first opening the dataset when no explicit timestamp is provided. Note the user can still change the data in the xarray dataset before it is loaded by explicitly writing data to a timestamp earlier than when it is opened. To mimic the earlier behavior, a user can explicitly open the TileDB array for a future timestamp.

Fixes:
* Uses saved timestamp when getting TileDB metadata

Changes:
* If no timestamp is provided, use the timestamp at time of opening.